Split logic into multiple files, adapt merge3 and incorporate PR feedback
